# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "This is API documentation on DataBx Semeru2 Microservices",
    "version": "1.0.0",
    "title": "DataBx Semeru2 Microservices",
    "contact": {
      "email": "akbar.muhammadakbarmaulana@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "tags": [
    {
      "name": "credential",
      "description": "store information of account_id, email_address, password and salt"
    }
  ],
  "paths": {
    "/select/credential/id": {
      "post": {
        "tags": [
          "credential"
        ],
        "summary": "Get single credential by given id",
        "description": "Should only return single account",
        "operationId": "GetCredentialById",
        "security": [
          {"basicAuth":  []}
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "$ref": "#/components/parameters/Host"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Credential"
        },
        "responses": {
          "200": {
            "description": "GetCredentialById success, with given request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Credential"
                }
              }
            }
          },
          "400": {
            "description": "Check request, body not comply with request rules"
          },
          "401": {
            "description": "You are not permitted!"
          },
          "500": {
            "description": "General error, system cannot tell you what wrongs"
          }
        }
      }
    },
    "/select/credential/email": {
      "post": {
        "tags": [
          "credential"
        ],
        "summary": "Get single credential by given email address",
        "description": "Should only return single account",
        "operationId": "GetCredentialByEmail",
        "security": [
          {"basicAuth":  []}
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "$ref": "#/components/parameters/Host"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Credential"
        },
        "responses": {
          "200": {
            "description": "GetCredentialById success, with given request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Credential"
                }
              }
            }
          },
          "400": {
            "description": "Check request, body not comply with request rules"
          },
          "401": {
            "description": "You are not permitted!"
          },
          "500": {
            "description": "General error, system cannot tell you what wrongs"
          }
        }
      }
    },
    "/insert/credential": {
      "post": {
        "tags": [
          "credential"
        ],
        "summary": "Insert credential by given email and password",
        "description": "Only create single account",
        "operationId": "InsertCredential",
        "security": [
          {"basicAuth":  []}
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Host"
          },
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Credential"
        },
        "responses": {
          "200": {
            "description": "InsertCredential success, with given request"
          },
          "400": {
            "description": "Check request, body not comply with request rules"
          },
          "401": {
            "description": "You are not permitted!"
          },
          "500": {
            "description": "General error, system cannot tell you what wrongs"
          }
        }
      }
    },
    "/update/credential": {
      "post": {
        "tags": [
          "credential"
        ],
        "summary": "Update credential by given account_id, email and password",
        "description": "Only update single account",
        "operationId": "UpdateCredential",
        "security": [
          {"basicAuth":  []}
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Host"
          },
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Credential"
        },
        "responses": {
          "200": {
            "description": "UpdateCredential success, with given request"
          },
          "400": {
            "description": "Check request, body not comply with request rules"
          },
          "401": {
            "description": "You are not permitted!"
          },
          "500": {
            "description": "General error, system cannot tell you what wrongs"
          }
        }
      }
    },
    "/delete/credential": {
      "post": {
        "tags": [
          "credential"
        ],
        "summary": "Delete credential by given account_id",
        "description": "Only delete single account",
        "operationId": "DeleteCredential",
        "security": [
          {"basicAuth":  []}
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Host"
          },
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Credential"
        },
        "responses": {
          "200": {
            "description": "DeleteCredential success, with given request"
          },
          "400": {
            "description": "Check request, body not comply with request rules"
          },
          "401": {
            "description": "You are not permitted!"
          },
          "500": {
            "description": "General error, system cannot tell you what wrongs"
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:7718/semeru2/api/v1"
    }
  ],
  "components": {
    "requestBodies": {
      "Credential": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Credential"
            }
          }
        }
      }
    },
    "schemas": {
      "Credential": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "account_id": {
            "type": "integer",
            "format": "uint",
            "minimum": 1
          },
          "email_address": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "salt": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    },
    "parameters": {
      "Host": {
        "name": "host",
        "in": "header",
        "description": "Id that defines you as host",
        "required": false,
        "schema": {
          "type": "string",
          "maxLength": 5
        }
      },
      "RequestId": {
        "name": "request_id",
        "in": "header",
        "description": "Request id that accompanies, what else request",
        "required": true,
        "schema": {
          "type": "string",
          "maxLength": 36,
          "pattern": "[A-Za-z0-9]{8}-[A-Za-z0-9]{4}-[A-Za-z0-9]{4}-[A-Za-z0-9]{4}-[A-Za-z0-9]{12}"
        }
      }
    }
  },
  "security": [
    {
      "basicAuth": []
    }
  ]
}